dry protocol no longer tries to fake the relevant git command
This wasn't correct on ssh connections. protocol_canonicalize_path makes
it more correct, but it'd still be wrong for git init. It's just easier
to not try than it is to devise weird ways to fake it.

We might do a protocols/ssh.dry, local.dry, etc. so we can get pretty
log output without complicating lots of things